Provisioning: move ref validator into git package next to branch validator
`IsValidRef` and the commit-hash regex now live alongside
`IsValidGitBranchName` in `apps/provisioning/pkg/repository/git`,
removing the duplicated branch-name rules from the `repository` package.
`ErrInvalidRef` stays in `repository` with the other sentinels.
