+++ pkg/registry/apis/provisioning/files.go
@@ -17,6 +17,7 @@ import (
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
 	"github.com/grafana/grafana/apps/provisioning/pkg/quotas"
 	"github.com/grafana/grafana/apps/provisioning/pkg/repository"
+	"github.com/grafana/grafana/apps/provisioning/pkg/repository/git"
 	"github.com/grafana/grafana/apps/provisioning/pkg/safepath"
 	"github.com/grafana/grafana/pkg/apimachinery/apis/common/v0alpha1"
 	"github.com/grafana/grafana/pkg/apimachinery/utils"
@@ -199,7 +200,7 @@ func (c *filesConnector) parseRequestOptions(r *http.Request, name string, repo 
 
 	// Reject unvalidated refs before they reach any backend. Empty is allowed and
 	// is defaulted to the configured branch downstream.
-	if !repository.IsValidRef(opts.Ref) {
+	if !git.IsValidRef(opts.Ref) {
 		return opts, repository.ErrInvalidRef
 	}
